Added webpack tool chain for es syntax support
Heroku will not execute babel transform during deploy. Therefore,
webpack was implemented to handle es syntax and Heroku build platform
requirements.

Webpack was installed with necessary dependencies to support es syntax.
Note: `async await` is not supported by expected config and [this
solution was
implemented](https://dev.to/btandayamo/a-guide-to-through-async-await-with-babel-and-webpack-mh5)