Add chained npm start script
Due to the dyno config for Heroku the only available build command is
`npm start`. Working within this the `npm start` script chained with
the custom `npm run build` and `npm run start:build` this should work
with the dyno build command.

The objective of this change to was enable a more streamline deploy
within Heroku's build platform constraints.